# docs/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "GID Task Flow API"
  version: "0.1.0"
  description: "REST contract for GID Task Flow, a single-user task and time-tracking app. Request bodies are generated from the zod schemas in `src/lib/validators.ts`; response shapes mirror the routes' actual payloads. All endpoints except `/api/health` and `/api/auth/login` require the `stl_session` cookie. Note: generated object schemas show `additionalProperties: false`, but the routes use zod's default object behaviour, which ignores unknown keys instead of rejecting them. This file is generated — do not edit by hand; run `npm run docs:openapi`."
  license:
    name: "MIT"
servers:
  - url: "/"
    description: "Same-origin app (Next.js App Router route handlers)"
...
security:
  - cookieAuth: []
paths:
...
  "/api/breaks/log":
    post:
      tags:
        - "Breaks"
      summary: "Log a finished break"
      description: "The client states which break was taken and when it started; the task and its completion are written in one server-side transaction, so a break can never be half-logged (UX-03/FL-02)."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/BreakLogInput"
      responses:
        "201":
          description: "Logged."
          content:
            "application/json":
              schema:
                type: "object"
                properties:
                  success:
                    type: "boolean"
                  taskId:
                    type: "integer"
                  minutes:
                    type: "integer"
        "400":
          description: "Payload rejected by zod validation or malformed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "401":
          description: "Not authenticated."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Job or break type not found."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
...
  "/api/settings":
    get:
      tags:
        - "Settings"
      summary: "Ensure a settings row exists (health of auth setup)"
      description: "Creates the singleton `UserSettings` row when missing and returns `{ ok: true }`. Work schedules are per-job and live on `/api/jobs/{jobId}`; the POST form of this route is a kept-for-compatibility no-op."
      responses:
        "200":
          description: "Settings row present."
          content:
            "application/json":
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"
                    enum:
                      - true
                required:
                  - "ok"
        "401":
          description: "Not authenticated."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    post:
      tags:
        - "Settings"
      summary: "No-op compatibility endpoint"
      description: "Authenticates and returns `{ ok: true }`; it writes nothing."
      responses:
        "200":
          description: "No-op."
          content:
            "application/json":
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"
                    enum:
                      - true
                required:
                  - "ok"
        "401":
          description: "Not authenticated."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    patch:
      tags:
        - "Settings"
      summary: "Change the password"
      description: "Verifies the current password, re-hashes at bcrypt cost 12 and bumps `tokenVersion`, revoking every existing session. The new `stl_session` cookie is returned in the same response so the user stays logged in."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: "Password changed."
          content:
            "application/json":
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"
                    enum:
                      - true
                required:
                  - "ok"
        "400":
          description: "Payload rejected by zod validation or malformed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "401":
          description: "Current password is incorrect."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unable to change password."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: "apiKey"
      in: "cookie"
      name: "stl_session"
      description: "HttpOnly JWT session cookie issued by `/api/auth/login` (7-day max age, `SameSite=Lax`, `Secure` in production). Carries `{ sub, tv }`; `tv` must match `UserSettings.tokenVersion` so a password change or logout revokes old tokens (SEC-06). There is no CSRF token and no Origin check on mutating routes — see MF-02/AI-02 in the audit for the residual risk."
  schemas:
    Error:
      type: "object"
      description: "Every error body has `error`. Validation failures additionally return `issues` (zod) or `details` (export query)."
      properties:
        error:
          type: "string"
        issues:
          type: "array"
          description: "zod issues, present on some 400 responses."
          items:
            type: "object"
            properties:
              path:
                type: "array"
                items:
                  type: "string"
              message:
                type: "string"
            required:
              - "message"
      required:
        - "error"
...
    ChangePasswordInput:
      type: "object"
      properties:
        currentPassword:
          type: "string"
          minLength: 1
        newPassword:
          type: "string"
          minLength: 6
          maxLength: 128
        confirmPassword:
          type: "string"
          minLength: 6
          maxLength: 128
      required:
        - "currentPassword"
        - "newPassword"
        - "confirmPassword"
      additionalProperties: false
...
    BreakLogInput:
      type: "object"
      properties:
        jobId:
          type: "integer"
          exclusiveMinimum: 0
          maximum: 9007199254740991
        projectId:
          type: "integer"
          exclusiveMinimum: 0
          maximum: 9007199254740991
        name:
          type: "string"
          minLength: 1
          maxLength: 100
        startedAt:
          type: "string"
          format: "date-time"
          pattern: "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
      required:
        - "jobId"
        - "name"
        - "startedAt"
      additionalProperties: false
...